adds Nightwatch + Saucelabs documentation to Readme & simplify tests
